requirements.txt
Flask==2.3.3
gspread==5.11.3
google-auth==2.22.0
pyTelegramBotAPI==4.13.0
